Moved Metods into SceneObjectExtensions
