+++ Source/Core/Runtime/SceneObjects/SceneObjectExtensions.cs
@@ -14,7 +14,7 @@ using Object = UnityEngine.Object;
 namespace VRBuilder.Core.SceneObjects
 {
     /// <summary>
-    /// Helper class that adds functionality to any <see cref="ISceneObject"/>.
+    /// Helper class that adds functionality to any <see cref="ISceneObject"/> and some utility functions.
     /// </summary>
     public static class SceneObjectExtensions
     {
@@ -125,6 +125,100 @@ namespace VRBuilder.Core.SceneObjects
 
             IEnumerable<Type> typesToIgnore = GetTypesFromComponents(excludedFromBeingRemoved);
             RemoveProperty(sceneObject, processProperty, removeDependencies, typesToIgnore);
+        }
+
+
+        /// <summary>
+        /// Automatically sets up a scene object by ensuring it has a <see cref="ProcessSceneObject"/> component
+        /// and adding a process property of the specified type.
+        /// </summary> <param name="selectedSceneObject"> The <see cref="GameObject"/> to be set up as a scene object. </param>
+        /// <param name="valueType"> The type of the process property to add to the scene object. </param>
+        /// <remarks>
+        /// The method will attempt to find an implementation of this type with a default attribute. 
+        /// If none is found, it will use the first found implementation without a default attribute.
+        /// </remarks>
+        public static void SceneObjectAutomaticSetup(GameObject selectedSceneObject, Type valueType)
+        {
+            ISceneObject sceneObject = selectedSceneObject.GetComponent<ProcessSceneObject>() ?? selectedSceneObject.AddComponent<ProcessSceneObject>();
+            Type concreteTypeToAdd = GetImplementation(valueType);
+
+            if (concreteTypeToAdd != null)
+            {
+                sceneObject.AddProcessProperty(concreteTypeToAdd);
+            }
+        }
+
+        /// <summary>
+        /// Retrieves the concrete implementation type for the specified value type.
+        /// </summary>
+        /// <param name="valueType">The type for which to find a concrete implementation.</param>
+        /// <returns>
+        /// The concrete implementation type for the specified value type, or <c>null</c>
+        /// if no suitable implementation is found.
+        /// </returns>
+        /// remarks>
+        /// Attempts to find an implementation marked with a default attribute first.
+        /// If no such implementation is found, it falls back to finding an implementation without the default attribute.
+        /// </remarks>
+        private static Type GetImplementation(Type valueType)
+        {
+            Type concreteTypeToAdd = ReflectionUtils.GetImplementationWithDefaultAttribute(valueType);
+
+            if (concreteTypeToAdd == null)
+            {
+                concreteTypeToAdd = ReflectionUtils.GetImplementationWithoutDefaultAttribute(valueType);
+            }
+
+            return concreteTypeToAdd;
+        }
+
+        /// <summary>
+        /// Performs an automatic undo setup for a scene object.
+        /// </summary>
+        /// <param name="selectedSceneObject">The GameObject to process.</param>
+        /// <param name="valueType">The type used for reflection.</param>
+        /// <param name="alreadyAttachedProperties">Array of components that are considered original.</param>
+        public static bool UndoSceneObjectAutomaticSetup(GameObject selectedSceneObject, Type valueType, Component[] alreadyAttachedProperties)
+        {
+            var sceneObject = selectedSceneObject.GetComponent<ProcessSceneObject>();
+            if (sceneObject == null)
+            {
+                return false;
+            }
+
+            RemoveProcessProperty(sceneObject, valueType, alreadyAttachedProperties);
+
+            List<Component> sortedComponents = ComponentUtils.GetSortedNonOriginalComponents(selectedSceneObject, alreadyAttachedProperties);
+
+            // Remove components in reverse topological order so that dependents are removed before their dependencies.
+            for (int i = sortedComponents.Count - 1; i >= 0; i--)
+            {
+                Object.DestroyImmediate(sortedComponents[i]);
+            }
+
+            return true;
+        }
+
+        /// <summary>
+        /// Removes a process property and from the scene object based on the provided type.
+        /// </summary>
+        /// <param name="sceneObject">The scene object.</param>
+        /// <param name="valueType">The type for determining the property to remove.</param>
+        /// <param name="alreadyAttachedProperties">Array of components that are considered original.</param>
+        /// <remarks>
+        /// This method first attempts to find a concrete type with a default attribute. If none is found, it uses the first found implementation without a default attribute.
+        /// <remarks>
+        public static void RemoveProcessProperty(ISceneObject sceneObject, Type valueType, Component[] alreadyAttachedProperties)
+        {
+            Type concreteTypeToRemove = ReflectionUtils.GetImplementationWithDefaultAttribute(valueType);
+            if (concreteTypeToRemove == null)
+            {
+                concreteTypeToRemove = ReflectionUtils.GetImplementationWithoutDefaultAttribute(valueType);
+            }
+            if (concreteTypeToRemove != null)
+            {
+                sceneObject.RemoveProcessProperty(concreteTypeToRemove, true, alreadyAttachedProperties);
+            }
         }
 
         private static void RemoveProperty(ISceneObject sceneObject, Type typeToRemove, bool removeDependencies, IEnumerable<Type> typesToIgnore)
